Rename Sha1 -> SHA1
