+++ library/sha1/api/sha1.api
@@ -1,4 +1,4 @@
-public final class org/kotlincrypto/hash/Sha1 : org/kotlincrypto/core/Digest {
+public final class org/kotlincrypto/hash/SHA1 : org/kotlincrypto/core/Digest {
 	public fun <init> ()V
 }
