+++ library/sha1/src/commonTest/kotlin/org/kotlincrypto/hash/SHA1UnitTest.kt
@@ -18,8 +18,8 @@ package org.kotlincrypto.hash
 import org.kotlincrypto.core.Digest
 import kotlin.test.Test
 
-open class Sha1UnitTest: DigestUnitTest() {
-    override val digest: Digest = Sha1()
+open class SHA1UnitTest: DigestUnitTest() {
+    override val digest: Digest = SHA1()
     final override val expectedResetHash: String = "da39a3ee5e6b4b0d3255bfef95601890afd80709"
     final override val expectedUpdateSmallHash: String = "3e87f2fb5366045ef4fcaf3a845554d16b36f69d"
     final override val expectedUpdateMediumHash: String = "c0e5c75fba36a5a24ad475ac2321c581aea8008e"
